Trim long finding descriptions in the customer report
This is so that the descriptions don't overflow the page (LaTeX table cells cannot span pages, as far as I know) or cause other issues. For more information, see:
 - https://github.com/cisagov/cyhy-reports/issues/124
 - https://github.com/cisagov/cyhy-reports/issues/123

NOTE: I wanted to make the cutoff 4096 characters, but that barely fit on a single page when there was a single affected host.  I went with 3072 instead to save space on the page for when there are many affected hosts listed.